refactor(chat): replace history sheet with inline view switcher
Simplify the history UX based on feedback that the sheet+dropdown
pattern was overkill. The chat panel now toggles between a chat view
and a history view via a header bar:

- Chat view header shows "新しいチャット" and "履歴" buttons.
- Pressing 履歴 swaps the body for a flat thread list with a back arrow
  in the header.
- Selecting a thread loads it and returns to the chat view.
- Delete is a small trash icon shown on row hover, confirmed via
  AlertDialog. The Sheet and DropdownMenu components have been removed
  along with the standalone history sheet file.
